CepespData/requirements.txt
awsebcli==3.14.1
Babel==2.5.3
boto3==1.7.24
botocore==1.10.24
cached-property==1.4.3
cachetools==2.1.0
cement==2.8.2
click==6.7
colorama==0.3.9
docker==3.4.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Flask==1.0.2
Flask-Babel==0.11.2
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
httplib2==0.11.3
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.0
nose==1.3.7
numpy==1.14.3
pandas==0.22.0
pathspec==0.5.5
pyasn1==0.4.3
pyasn1-modules==0.2.2
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
PyYAML==3.12
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
uritemplate==3.0.0
websocket-client==0.47.0
Werkzeug==0.14.1
s3fs==0.1.6
